+++ src/fluxions_gte_image.cpp
@@ -643,7 +643,7 @@ namespace Fluxions {
 
 		int imageColorRange = 255;
 		if (std::is_integral_v<scalar_type>) {
-			imageColorRange = std::max<int>(maxColorFound, 255);
+			imageColorRange = std::max<int>((int)maxColorFound, 255);
 		}
 		else if (std::is_floating_point_v<scalar_type>) {
 			imageColorRange = std::min<int>(65535, (int)std::floor(std::max(maxColorFound, 1.0f) * 255.99f));
